refactor: 비슷한 훅끼리 한 폴더 내로 구조화
